backend/requirements.txt
# 하나 더 (Hana More) - Backend 의존성 / Python 3.11 기준

# 웹 프레임워크
fastapi==0.115.5
uvicorn[standard]==0.32.1

# 인증 / 보안
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.12

# 환경변수
python-dotenv==1.0.1

# Supabase 클라이언트
supabase==2.10.0

# 데이터 검증
pydantic==2.10.3
pydantic-settings==2.6.1

# ML / 스코어링 엔진
xgboost==2.1.3
shap==0.46.0
scikit-learn==1.5.2
numpy==1.26.4
pandas==2.2.3

# 수입 예측
prophet==1.1.6

# 해싱 / 암호화
pysha3==1.0.2
pycryptodome==3.21.0

# 비동기 HTTP 클라이언트
httpx==0.27.2

# 배치 작업
celery==5.4.0
redis==5.2.0

# 유틸리티
python-dateutil==2.9.0
pytz==2024.2
